Change the input layer order to match the GIMP plug-in
The 8bf plug-in sends layers in bottom to top order, whereas the
G'MIC-Qt plug-in for GIMP sends layers in top to bottom order.
We now reverse the layer list to match the behavior of the G'MIC-Qt
plug-in for GIMP.

Also fixed the ActiveAnd* input cases to match the GIMP plug-in, in
those modes it only uses the first layer that is above or below the
active layer.