Use the GMIC 8bf image format when writing the output images
This allows a planar image with tiling to be written for the images
that are copied into the active layer of the plugin host.
The planar with tiling image format reduces memory consumption when
loading most images into the plugin host.

The images that are saved to an external file will be written in an
interleaved pixel format, and the plugin will convert them to a
different format when it saves the output image.

Writing the GMIC 8bf image format also allows the 8bf plugin to be
extended with support for 32-bit floating point images, which was not
possible when using PNG as the output format.

The Gmic8bfLayer structure now uses a cimg_library::CImg<float> as the
image data format, this removes the need to convert a QImage to the
G"MIC native format.
Added some code to automatically launch the Visual Studio debugger when
running a debug build.